Moving Map component
The map component has been moved to the route directory in order to use it as the landing page. The earlier approach was causing some rendering issues such as flickering
