+++ FindMyClass/app/index.jsx
@@ -1,32 +1,23 @@
-import React, { useEffect, useState } from 'react';
-import { View, ActivityIndicator } from 'react-native';
-import { useRouter, useSegments } from 'expo-router';
-import "react-native-get-random-values";
+import React from 'react';
+import { View, StyleSheet } from 'react-native';
+import { useRoute } from '@react-navigation/native'; 
+import ToggleCampusMap from '../components/ToggleCampusMap';
+import FloatingChatButton from '../components/FloatingChatButton';
 
+export default function MapScreen() {  // ✅ Renamed to avoid conflict
+  const route = useRoute();
+  const searchText = route?.params?.searchText || ''; // Cleaner fallback
 
-export default function Home() {
-
-  const router = useRouter();
-  const [isReady, setIsReady] = useState(false);
-
-  useEffect( () =>  {
-    let timeOut = setTimeout( () => {
-      setIsReady(true);
-      router.replace("/screens/map");
-    }, 100);
-    return () => clearTimeout(timeOut);
-  },[]);
-
-if (!isReady){
-  return   (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <ActivityIndicator 
-        testID="loading-indicator"
-        size="large" 
-        color="#800000" 
-      />
+  return (
+    <View style={styles.container} testID="map-container">
+      <ToggleCampusMap searchText={searchText} testID="toggle-campus-map" />
+      <FloatingChatButton testID="floating-chat-button" />
     </View>
-  );}
-  return null;
+  );
 }
 
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
