--- FindMyClass/app/screens/map.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { View, StyleSheet } from 'react-native';
-import { useRoute } from '@react-navigation/native'; 
-import ToggleCampusMap from '../../components/ToggleCampusMap';
-import FloatingChatButton from '../../components/FloatingChatButton';
-
-export default function MapScreen() {  // ✅ Renamed to avoid conflict
-  const route = useRoute();
-  const searchText = route?.params?.searchText || ''; // Cleaner fallback
-
-  return (
-    <View style={styles.container} testID="map-container">
-      <ToggleCampusMap searchText={searchText} testID="toggle-campus-map" />
-      <FloatingChatButton testID="floating-chat-button" />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
